+++ projects/angular2-smart-table/src/lib/components/cell/cell-view-mode/custom-view.component.ts
@@ -1,7 +1,6 @@
 import {Component, Input, OnDestroy, OnInit, ViewChild, ViewContainerRef,} from '@angular/core';
 
 import {Cell} from '../../../lib/data-set/cell';
-import {ViewCell} from './view-cell';
 
 @Component({
   selector: 'custom-view-component',
@@ -18,22 +17,16 @@ export class CustomViewComponent implements OnInit, OnDestroy {
   ngOnInit() {
     if (this.cell && !this.customComponent) {
       this.customComponent = this.dynamicTarget.createComponent(this.cell.getColumn().renderComponent);
-      Object.assign(this.customComponent.instance, this.getPatch());
-      const onComponentInitFunction = this.cell.getColumn().onComponentInitFunction;
-      if (onComponentInitFunction !== undefined) {
-        onComponentInitFunction(this.customComponent.instance, this.getPatch());
+      const componentInitFunction = this.cell.getColumn().componentInitFunction;
+      if (componentInitFunction === undefined) {
+        console.error('Since version 3.0.0, a custom renderer always needs a componentInitFunction');
+      } else {
+        componentInitFunction(this.customComponent.instance, this.cell);
       }
     }
   }
 
   ngOnDestroy() {
     this.customComponent.destroy();
-  }
-
-  protected getPatch(): ViewCell {
-    return {
-      value: this.cell.getValue(),
-      rowData: this.cell.getRow().getData()
-    }
   }
 }