+++ projects/angular2-smart-table/src/public-api.ts
@@ -3,7 +3,6 @@
  */
 
 export * from './lib/angular2-smart-table.module';
-export { ViewCell } from './lib/components/cell/cell-view-mode/view-cell';
 export { DefaultEditor, Editor } from './lib/components/cell/cell-editors/default-editor';
 export { DefaultFilter, Filter } from './lib/components/filter/filter-types/default-filter'
 export { Cell } from './lib/lib/data-set/cell';